requirements.txt
django==2.2.5
psycopg2
pytest
pytest-django
djangorestframework
lxml
ConfigArgParse
requests
